+++ CHANGELOG.md
@@ -1,13 +1,37 @@
-# Bashpack changes
+# Changes
 
 The released versions changelogs below are only about the main repository.
 
+### 3.0.0
+*2025/02/21*
+
+##### Added
+- The CLI can now be installed offline by itself
+- Anyone can now host a repository for its own subcommands (edit sourceslist option from the configuration directory) 
+- Anyone can now host a repository for its own CLI (edit cli_url option from the configuration file)
+- Repositories are compatibles with Github, but can also be simple directories listing web servers
+- Documentation website at https://bashpack-project.github.io
+- Subcommand template is available with the option "-n'
+
+##### Modified
+- The CLI has been redesigned has a "script manager", to download/install/run/remove scripts as subcommands
+- The CLI helper can now easily create an automation from a simple function that allow to automatically run a subcommand on the system (systemd only for now)
+- When creating subcommands, the function "init_command()" can be used to run any requirements on the system (create automation, file etc... or anything else)
+- Subcommands provided from this repository has moved to the [commands repository](https://github.com/bashpack-project/commands)
+- Subcommands "update" and "install" are now merged in the "package" subcommand
+- Subcommand "verify" option "-c/--commands" replaced by "-d/--dependencies"
+- Publication system has been removed and replaced by the configuration option "cli_url"
+
+##### Fixed
+- Subcommand "firewall" restoration option wasn't able to read the given file name
+
+------------
 ### 2.0.0
 *2024/11/19*
 ##### Added
 - Log support in /var/log/bashpack/ and "--get-logs" option
-- Sub command "update" support of DNF (and YUM fallback)
-- Sub command "firewall" for firewall management (nftables only for now)
+- Subcommand "update" support of DNF (and YUM fallback)
+- Subcommand "firewall" for firewall management (nftables only for now)
 - This changelog file :D
 
 ##### Modified
@@ -16,22 +40,22 @@ The released versions changelogs below are only about the main repository.
 - Self update can now be forced with "-u --force" option 
 - Self installation improved by automatically detect the better $PATH
 - Self installation can now read a given publication with "-i publication_name" 
-- Sub command "verify" is now more used in installation process
-- Sub command "verify" can now detect presence of required commands
+- Subcommand "verify" is now more used in installation process
+- Subcommand "verify" can now detect presence of required commands
 - Config file renamed from bashpack_config to bashpack.conf
 - Detection of systemd improved to be able to install on compatible systems only
 - Detection of pkg-config improved to be able to install on compatible systems only
 - Moving sources from /usr/local/src/ to /opt/
 
 ##### Removed
-- Sub command "verify" download/extraction
+- Subcommand "verify" download/extraction
 
 ------------
 ### 1.1.1
 *2024/10/06*
 ##### Added
-- Sub command "verify" now replaces command option "-t"
-- Sub command "verify" can now test repository reachability and test releases download/extraction
+- Subcommand "verify" now replaces command option "-t"
+- Subcommand "verify" can now test repository reachability and test releases download/extraction
 
 ##### Modified
 - Self update is now taking in account the configuration file and the user custom values